Move draw functions using Context API
