+++ src/components/App.js
@@ -3,12 +3,11 @@ import CanvasContainer from './CanvasContainer'
 import MusicContainer from './MusicContainer'
 import { Route } from 'react-router-dom';
 
-
 function App() {
   return (
     <div className="App">
       {/* <MusicContainer /> */}
-      <CanvasContainer />
+        <CanvasContainer />
     </div>
   );
 }
