+++ src/components/Canvas.js
@@ -1,54 +1,14 @@
-import React, { useState, useRef, useEffect }from 'react'
+import React, { useContext, useEffect }from 'react'
 import '../styles/Canvas.scss'
+import { useCanvas, CanvasContext } from '../context/CanvasContext'
 
 
 const Canvas = () => {
-  const canvasRef = useRef(null);
-  const contextRef = useRef(null);
-  const [isDrawing, setIsDrawing] = useState(false)
+ const { prepareCanvas, contextRef, canvasRef, startDrawing, finishDrawing, draw } = useCanvas();
 
-  useEffect(() => {
-    const canvas = canvasRef.current
-    canvas.width = window.innerWidth * 2;
-    canvas.height = window.innerHeight * 2;
-    canvas.style.width = `${window.innerWidth}px`;
-    canvas.style.height = `${window.innerHeight}px`;
-
-    const context = canvas.getContext("2d")
-    context.scale(2, 2);
-    context.lineCap = "round";
-    context.strokeStyle = "black";
-    context.lineWidth = 5;
-    contextRef.current = context;
-  }, [])
-
-  const startDrawing = ({ nativeEvent }) => {
-    const { offsetX, offsetY } = nativeEvent;
-    contextRef.current.beginPath();
-    contextRef.current.moveTo(offsetX, offsetY);
-    setIsDrawing(true);
-  };
-
-  const finishDrawing = () => {
-    contextRef.current.closePath();
-    setIsDrawing(false);
-  };
-
-  const draw = ({ nativeEvent }) => {
-    if (!isDrawing) {
-      return;
-    }
-    const { offsetX, offsetY } = nativeEvent;
-    contextRef.current.lineTo(offsetX, offsetY);
-    contextRef.current.stroke();
-  };
-
-   const clearCanvas = () => {
-    const canvas = canvasRef.current;
-    const context = canvas.getContext("2d")
-    context.fillStyle = "white"
-    context.fillRect(0, 0, canvas.width, canvas.height)
-  }
+ useEffect(() => {
+   prepareCanvas();
+ }, [])
 
   return (
    <canvas
