+++ src/index.js
@@ -3,7 +3,11 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './components/App';
 import { BrowserRouter } from 'react-router-dom';
+import { CanvasContextProvider } from "./context/CanvasContext";
+
 
 ReactDOM.render(
-    <App />, document.getElementById('root')
+    <CanvasContextProvider>
+        <App />
+    </CanvasContextProvider>, document.getElementById('root')
 );
